starting to break out mining
